chore: move entity classes into dedicated package
