+++ src/main/java/com/codenames/codenames/backend/database/repository/PlayerRepository.java
@@ -1,6 +1,8 @@
 package com.codenames.codenames.backend.database.repository;
 
 import com.codenames.codenames.backend.database.entity.Player;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
